fix some bugs, fix circular imports
