+++ src/motion_planning/scripts/competition_task.py
@@ -1,10 +1,9 @@
 from template_tasks import *
 
 GateTask = ListTask([
-					# Submerge
-					MoveToPoseLocalTask(0, 0, -2, 0, 0, 0),
-					# Go straight
-					MoveToPoseLocalTask(10, 0, 0, 0, 0, 0)
-					])
-
+	# Submerge
+	MoveToPoseLocalTask(0, 0, -2, 0, 0, 0),
+	# Go straight
+	MoveToPoseLocalTask(10, 0, 0, 0, 0, 0)
+	])
 CompetitionTask = ListTask([GateTask])
