+++ src/motion_planning/scripts/template_tasks.py
@@ -3,22 +3,22 @@ from geometry_msgs.msg import Pose, Quaternion
 from tf.transformations import quaternion_from_euler
 import tf
 import task_utils
-from task_runner import TaskRunner
+from task_state import TaskState
 
 class MoveToPoseGlobalTask(Task):
     """Move to pose given in global coordinates."""
 
     def __init__(self, x, y, z, roll, pitch, yaw):
     	self.desired_pose = Pose()
-    	self.desired_pose.point.x = x
-    	self.desired_pose.point.y = y
-    	self.desired_pose.point.z = z
-    	self.desired_pose.quaternion = Quaternion(quaternion_from_euler(roll, pitch, yaw))
+    	self.desired_pose.position.x = x
+    	self.desired_pose.position.y = y
+    	self.desired_pose.position.z = z
+    	self.desired_pose.orientation = Quaternion(*quaternion_from_euler(roll, pitch, yaw))
         super(MoveToPoseGlobalTask, self).__init__()
 
     def _task_run(self):
     	task_utils.publish_desired_pose_global(self.desired_pose)
-        if(task_utils.at_pose(self.desired_pose, TaskRunner.STATE.pose)):
+        if(task_utils.at_pose(self.desired_pose, TaskState.STATE.pose)):
             self.finish()
 
 class MoveToPoseLocalTask(MoveToPoseGlobalTask):
@@ -26,7 +26,7 @@ class MoveToPoseLocalTask(MoveToPoseGlobalTask):
 
     def __init__(self, x, y, z, roll, pitch, yaw):
         super(MoveToPoseLocalTask, self).__init__(x, y, z, roll, pitch, yaw)
-        self.desired_pose = task_utils.transform(base_link, odom, pose=self.desired_pose)
+        self.desired_pose = task_utils.transform("base_link", "odom", pose=self.desired_pose)
 
 class HoldPositionTask(Task):
     """Hold position for a given number of seconds."""
